backend/app/requirements.txt
fastapi
sqlalchemy
aiogram
uvicorn
asyncpg
aiogram
dotenv
apscheduler
aiohttp-socks